renamed llms to llm_connectors and LLM to LLMConnector
