+++ ai_text_to_sql/llm_connectors/bard_llm.py
@@ -1,9 +1,9 @@
 from typing import Text
 
-from ai_text_to_sql.llms.llm import LLM
+from ai_text_to_sql.llm_connectors.llm_connector import LLMConnector
 
 
-class Bard(LLM):
+class Bard(LLMConnector):
     def create_prompt(self, user_input: Text, database_schema: Text) -> Text:
         pass
