+++ ai_text_to_sql/llm_connectors/llm_factory.py
@@ -2,8 +2,8 @@ import sys
 import inspect
 from typing import Optional, Dict, Type, Text
 
-from .llm import LLM
-from ai_text_to_sql.llms import *
+from .llm_connector import LLMConnector
+from ai_text_to_sql.llm_connectors import *
 
 
 class LLMFactory:
@@ -11,7 +11,7 @@ class LLMFactory:
     The class for building LLMs.
     """
     @staticmethod
-    def build_llm(name: Text, api_key: Text, **kwargs) -> LLM:
+    def build_llm(name: Text, api_key: Text, **kwargs) -> LLMConnector:
         """
         Build a LLM.
         :param name: The name of the LLM.
@@ -25,13 +25,13 @@ class LLMFactory:
             raise ValueError(f"Unsupported LLM: {name}")
 
     @staticmethod
-    def _discover_llms() -> Dict[str, Type[LLM]]:
+    def _discover_llms() -> Dict[str, Type[LLMConnector]]:
         """
         Discover available LLMs dynamically.
         :return: A dictionary mapping LLM names to their corresponding classes.
         """
         llms = {}
         for name, obj in inspect.getmembers(sys.modules[__name__]):
-            if inspect.isclass(obj) and issubclass(obj, LLM) and obj is not LLM:
+            if inspect.isclass(obj) and issubclass(obj, LLMConnector) and obj is not LLMConnector:
                 llms[obj.__name__] = obj
         return llms
